requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
google-cloud-storage==2.14.0
google-cloud-logging==3.8.0
google-cloud-error-reporting==1.1.3
google-cloud-monitoring==2.16.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
qrcode[pil]==7.4.2
reportlab==4.0.7
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.28.1
redis==5.0.1

